moves AttributionStrings to its own module
