+++ settings.gradle
@@ -1,3 +1,4 @@
+include ':strings'
 include ':feature:identity'
 include ':feature:subscriber-attributes'
 include ':utils'
