+++ strings/src/main/java/com/revenuecat/purchases/strings/AttributionStrings.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.subscriberattributes.logging
+package com.revenuecat.purchases.strings
 
 object AttributionStrings {
     const val NO_SUBSCRIBER_ATTRIBUTES_TO_SYNCHRONIZE = "No subscriber attributes to synchronize."
